Rename other non-hardcoded setup-ads or SetupAds to ads-onboarding.
